# PERF: skip libjoin fastpath for MultiIndex